update: UNIT3D linting
- github action updated with new ruleset in pint.json
- codebase linted with new ruleset
- contributors can now run `./vendor/bin/pint`
- action workflow will auto correct any lint issues upon commit/opened pull request

+++ app/Http/Controllers/GraveyardController.php
@@ -27,7 +27,7 @@ class GraveyardController extends Controller
      */
     public function index(): \Illuminate\Contracts\View\Factory|\Illuminate\View\View
     {
-        return \view('graveyard.index');
+        return view('graveyard.index');
     }
 
     /**
@@ -40,13 +40,13 @@ class GraveyardController extends Controller
         $resurrected = Graveyard::where('torrent_id', '=', $torrent->id)->first();
 
         if ($resurrected) {
-            return \to_route('graveyard.index')
-                ->withErrors(\trans('graveyard.resurrect-failed-pending'));
+            return to_route('graveyard.index')
+                ->withErrors(trans('graveyard.resurrect-failed-pending'));
         }
 
         if ($user->id === $torrent->user_id) {
-            return \to_route('graveyard.index')
-                ->withErrors(\trans('graveyard.resurrect-failed-own'));
+            return to_route('graveyard.index')
+                ->withErrors(trans('graveyard.resurrect-failed-own'));
         }
 
         $graveyard = new Graveyard();
@@ -54,21 +54,21 @@ class GraveyardController extends Controller
         $graveyard->torrent_id = $torrent->id;
         $graveyard->seedtime = $request->input('seedtime');
 
-        $v = \validator($graveyard->toArray(), [
+        $v = validator($graveyard->toArray(), [
             'user_id'    => 'required',
             'torrent_id' => 'required',
             'seedtime'   => 'required',
         ]);
 
         if ($v->fails()) {
-            return \to_route('graveyard.index')
+            return to_route('graveyard.index')
                 ->withErrors($v->errors());
         }
 
         $graveyard->save();
 
-        return \to_route('graveyard.index')
-            ->withSuccess(\trans('graveyard.resurrect-complete'));
+        return to_route('graveyard.index')
+            ->withSuccess(trans('graveyard.resurrect-complete'));
     }
 
     /**
@@ -82,10 +82,10 @@ class GraveyardController extends Controller
         $user = $request->user();
         $resurrection = Graveyard::findOrFail($id);
 
-        \abort_unless($user->group->is_modo || $user->id === $resurrection->user_id, 403);
+        abort_unless($user->group->is_modo || $user->id === $resurrection->user_id, 403);
         $resurrection->delete();
 
-        return \to_route('graveyard.index')
-            ->withSuccess(\trans('graveyard.resurrect-canceled'));
+        return to_route('graveyard.index')
+            ->withSuccess(trans('graveyard.resurrect-canceled'));
     }
 }
