+++ app/Http/Controllers/User/SeedboxController.php
@@ -31,11 +31,11 @@ class SeedboxController extends Controller
     {
         $user = User::where('username', '=', $username)->firstOrFail();
 
-        \abort_unless(($request->user()->group->is_modo || $request->user()->id == $user->id), 403);
+        abort_unless(($request->user()->group->is_modo || $request->user()->id == $user->id), 403);
 
         $seedboxes = Seedbox::where('user_id', '=', $user->id)->paginate(25);
 
-        return \view('user.seedbox.index', ['user' => $user, 'seedboxes' => $seedboxes]);
+        return view('user.seedbox.index', ['user' => $user, 'seedboxes' => $seedboxes]);
     }
 
     /**
@@ -50,15 +50,15 @@ class SeedboxController extends Controller
         $seedboxIps = $userSeedboxes->pluck('ip')->filter(fn ($ip) => filter_var($ip, FILTER_VALIDATE_IP));
         $seedboxNames = $userSeedboxes->pluck('name');
 
-        $v = \validator(
+        $v = validator(
             $request->input(),
             [
-                'name'  => [
+                'name' => [
                     'required',
                     'alpha_num',
                     Rule::notIn($seedboxNames),
                 ],
-                'ip'    => [
+                'ip' => [
                     'bail',
                     'required',
                     'ip',
@@ -72,7 +72,7 @@ class SeedboxController extends Controller
         );
 
         if ($v->fails()) {
-            return \to_route('seedboxes.index', ['username' => $user->username])
+            return to_route('seedboxes.index', ['username' => $user->username])
                 ->withErrors($v->errors());
         }
 
@@ -84,8 +84,8 @@ class SeedboxController extends Controller
         $seedbox->ip = $validated['ip'];
         $seedbox->save();
 
-        return \to_route('seedboxes.index', ['username' => $user->username])
-            ->withSuccess(\trans('user.seedbox-added-success'));
+        return to_route('seedboxes.index', ['username' => $user->username])
+            ->withSuccess(trans('user.seedbox-added-success'));
     }
 
     /**
@@ -98,11 +98,11 @@ class SeedboxController extends Controller
         $user = $request->user();
         $seedbox = Seedbox::findOrFail($id);
 
-        \abort_unless(($user->group->is_modo || $user->id == $seedbox->user_id), 403);
+        abort_unless(($user->group->is_modo || $user->id == $seedbox->user_id), 403);
 
         $seedbox->delete();
 
-        return \to_route('seedboxes.index', ['username' => $user->username])
-            ->withSuccess(\trans('user.seedbox-deleted-success'));
+        return to_route('seedboxes.index', ['username' => $user->username])
+            ->withSuccess(trans('user.seedbox-deleted-success'));
     }
 }
